move the rest of KV across
