+++ packages/wrangler/src/index.tsx
@@ -1,4 +1,3 @@
-import { StringDecoder } from "node:string_decoder";
 import TOML from "@iarna/toml";
 import chalk from "chalk";
 import supportsColor from "supports-color";
@@ -8,33 +7,15 @@ import { version as wranglerVersion } from "../package.json";
 import { readConfig } from "./config";
 import { d1 } from "./d1";
 import { devHandler, devOptions } from "./dev";
-import { confirm } from "./dialogs";
 import { workerNamespaceCommands } from "./dispatch-namespace";
 import { DeprecationError } from "./errors";
 import { generateHandler, generateOptions } from "./generate";
 import { initHandler, initOptions } from "./init";
-import { kvNamespace } from "./kv";
-import {
-	deleteKVBulkKeyValue,
-	deleteKVKeyValue,
-	getKVKeyValue,
-	getKVNamespaceId,
-	isKVKeyValue,
-	listKVNamespaceKeys,
-	putKVBulkKeyValue,
-	putKVKeyValue,
-	unexpectedKVKeyValueProps,
-} from "./kv/helpers";
+import { kvNamespace, kvKey, kvBulk } from "./kv";
 import { logger } from "./logger";
 import * as metrics from "./metrics";
 import { pages } from "./pages";
-import {
-	formatMessage,
-	ParseError,
-	parseJSON,
-	readFileSync,
-	readFileSyncToBuffer,
-} from "./parse";
+import { formatMessage, ParseError } from "./parse";
 import { previewHandler, previewOptions } from "./preview";
 import { publishOptions, publishHandler } from "./publish";
 import { pubSubCommands } from "./pubsub/pubsub-commands";
@@ -42,17 +23,10 @@ import { r2 } from "./r2";
 import { secret, secretBulkHandler, secretBulkOptions } from "./secret";
 import { tailOptions, tailHandler } from "./tail";
 import { updateCheck } from "./update-check";
-import {
-	listScopes,
-	login,
-	logout,
-	requireAuth,
-	validateScopeKeys,
-} from "./user";
+import { listScopes, login, logout, validateScopeKeys } from "./user";
 import { whoami } from "./whoami";
 
 import type { Config } from "./config";
-import type { KeyValue } from "./kv/helpers";
 import type Yargs from "yargs";
 
 export type ConfigPath = string | undefined;
@@ -452,479 +426,23 @@ function createCLIParser(argv: string[]) {
 			return kvNamespace(namespaceYargs.command(subHelp));
 		}
 	);
-
+	// kv
 	// :key
 	wrangler.command(
 		"kv:key",
 		"🔑 Individually manage Workers KV key-value pairs",
-		(kvKeyYargs) => {
-			return kvKeyYargs
-				.command(subHelp)
-				.command(
-					"put <key> [value]",
-					"Writes a single key/value pair to the given namespace.",
-					(yargs) => {
-						return yargs
-							.positional("key", {
-								type: "string",
-								describe: "The key to write to",
-								demandOption: true,
-							})
-							.positional("value", {
-								type: "string",
-								describe: "The value to write",
-							})
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The binding of the namespace to write to",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to write to",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							})
-							.option("ttl", {
-								type: "number",
-								describe: "Time for which the entries should be visible",
-							})
-							.option("expiration", {
-								type: "number",
-								describe:
-									"Time since the UNIX epoch after which the entry expires",
-							})
-							.option("metadata", {
-								type: "string",
-								describe: "Arbitrary JSON that is associated with a key",
-								coerce: (jsonStr: string): KeyValue["metadata"] => {
-									try {
-										return JSON.parse(jsonStr);
-									} catch (_) {}
-								},
-							})
-							.option("path", {
-								type: "string",
-								requiresArg: true,
-								describe: "Read value from the file at a given path",
-							})
-							.check(demandOneOfOption("value", "path"));
-					},
-					async ({ key, ttl, expiration, metadata, ...args }) => {
-						await printWranglerBanner();
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-						// One of `args.path` and `args.value` must be defined
-						const value = args.path
-							? readFileSyncToBuffer(args.path)
-							: // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-							  args.value!;
-
-						const metadataLog = metadata
-							? ` with metadata "${JSON.stringify(metadata)}"`
-							: "";
-
-						if (args.path) {
-							logger.log(
-								`Writing the contents of ${args.path} to the key "${key}" on namespace ${namespaceId}${metadataLog}.`
-							);
-						} else {
-							logger.log(
-								`Writing the value "${value}" to key "${key}" on namespace ${namespaceId}${metadataLog}.`
-							);
-						}
-
-						const accountId = await requireAuth(config);
-
-						await putKVKeyValue(accountId, namespaceId, {
-							key,
-							value,
-							expiration,
-							expiration_ttl: ttl,
-							metadata: metadata as KeyValue["metadata"],
-						});
-						await metrics.sendMetricsEvent("write kv key-value", {
-							sendMetrics: config.send_metrics,
-						});
-					}
-				)
-				.command(
-					"list",
-					"Outputs a list of all keys in a given namespace.",
-					(yargs) => {
-						return yargs
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to list",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to list",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								// In the case of listing keys we will default to non-preview mode
-								default: false,
-								describe: "Interact with a preview namespace",
-							})
-							.option("prefix", {
-								type: "string",
-								requiresArg: true,
-								describe: "A prefix to filter listed keys",
-							});
-					},
-					async ({ prefix, ...args }) => {
-						// TODO: support for limit+cursor (pagination)
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-
-						const accountId = await requireAuth(config);
-
-						const results = await listKVNamespaceKeys(
-							accountId,
-							namespaceId,
-							prefix
-						);
-						logger.log(JSON.stringify(results, undefined, 2));
-						await metrics.sendMetricsEvent("list kv keys", {
-							sendMetrics: config.send_metrics,
-						});
-					}
-				)
-				.command(
-					"get <key>",
-					"Reads a single value by key from the given namespace.",
-					(yargs) => {
-						return yargs
-							.positional("key", {
-								describe: "The key value to get.",
-								type: "string",
-								demandOption: true,
-							})
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to get from",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to get from",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							})
-							.option("preview", {
-								type: "boolean",
-								// In the case of getting key values we will default to non-preview mode
-								default: false,
-								describe: "Interact with a preview namespace",
-							})
-							.option("text", {
-								type: "boolean",
-								default: false,
-								describe: "Decode the returned value as a utf8 string",
-							});
-					},
-					async ({ key, ...args }) => {
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-
-						const accountId = await requireAuth(config);
-						const bufferKVValue = Buffer.from(
-							await getKVKeyValue(accountId, namespaceId, key)
-						);
-
-						if (args.text) {
-							const decoder = new StringDecoder("utf8");
-							logger.log(decoder.write(bufferKVValue));
-						} else {
-							process.stdout.write(bufferKVValue);
-						}
-						await metrics.sendMetricsEvent("read kv value", {
-							sendMetrics: config.send_metrics,
-						});
-					}
-				)
-				.command(
-					"delete <key>",
-					"Removes a single key value pair from the given namespace.",
-					(yargs) => {
-						return yargs
-							.positional("key", {
-								describe: "The key value to delete",
-								type: "string",
-								demandOption: true,
-							})
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to delete from",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to delete from",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							});
-					},
-					async ({ key, ...args }) => {
-						await printWranglerBanner();
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-
-						logger.log(
-							`Deleting the key "${key}" on namespace ${namespaceId}.`
-						);
-
-						const accountId = await requireAuth(config);
-
-						await deleteKVKeyValue(accountId, namespaceId, key);
-						await metrics.sendMetricsEvent("delete kv key-value", {
-							sendMetrics: config.send_metrics,
-						});
-					}
-				);
+		(keyYargs) => {
+			return kvKey(keyYargs.command(subHelp));
 		}
 	);
 
+	// kv
 	// :bulk
 	wrangler.command(
 		"kv:bulk",
 		"💪 Interact with multiple Workers KV key-value pairs at once",
-		(kvBulkYargs) => {
-			return kvBulkYargs
-				.command(subHelp)
-				.command(
-					"put <filename>",
-					"Upload multiple key-value pairs to a namespace",
-					(yargs) => {
-						return yargs
-							.positional("filename", {
-								describe: `The JSON file of key-value pairs to upload, in form [{"key":..., "value":...}"...]`,
-								type: "string",
-								demandOption: true,
-							})
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to insert values into",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to insert values into",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							});
-					},
-					async ({ filename, ...args }) => {
-						await printWranglerBanner();
-						// The simplest implementation I could think of.
-						// This could be made more efficient with a streaming parser/uploader
-						// but we'll do that in the future if needed.
-
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-						const content = parseJSON(readFileSync(filename), filename);
-
-						if (!Array.isArray(content)) {
-							throw new Error(
-								`Unexpected JSON input from "${filename}".\n` +
-									`Expected an array of key-value objects but got type "${typeof content}".`
-							);
-						}
-
-						const errors: string[] = [];
-						const warnings: string[] = [];
-						for (let i = 0; i < content.length; i++) {
-							const keyValue = content[i];
-							if (!isKVKeyValue(keyValue)) {
-								errors.push(
-									`The item at index ${i} is ${JSON.stringify(keyValue)}`
-								);
-							} else {
-								const props = unexpectedKVKeyValueProps(keyValue);
-								if (props.length > 0) {
-									warnings.push(
-										`The item at index ${i} contains unexpected properties: ${JSON.stringify(
-											props
-										)}.`
-									);
-								}
-							}
-						}
-						if (warnings.length > 0) {
-							logger.warn(
-								`Unexpected key-value properties in "${filename}".\n` +
-									warnings.join("\n")
-							);
-						}
-						if (errors.length > 0) {
-							throw new Error(
-								`Unexpected JSON input from "${filename}".\n` +
-									`Each item in the array should be an object that matches:\n\n` +
-									`interface KeyValue {\n` +
-									`  key: string;\n` +
-									`  value: string;\n` +
-									`  expiration?: number;\n` +
-									`  expiration_ttl?: number;\n` +
-									`  metadata?: object;\n` +
-									`  base64?: boolean;\n` +
-									`}\n\n` +
-									errors.join("\n")
-							);
-						}
-
-						const accountId = await requireAuth(config);
-						await putKVBulkKeyValue(accountId, namespaceId, content);
-						await metrics.sendMetricsEvent("write kv key-values (bulk)", {
-							sendMetrics: config.send_metrics,
-						});
-
-						logger.log("Success!");
-					}
-				)
-				.command(
-					"delete <filename>",
-					"Delete multiple key-value pairs from a namespace",
-					(yargs) => {
-						return yargs
-							.positional("filename", {
-								describe: `The JSON file of keys to delete, in the form ["key1", "key2", ...]`,
-								type: "string",
-								demandOption: true,
-							})
-							.option("binding", {
-								type: "string",
-								requiresArg: true,
-								describe: "The name of the namespace to delete from",
-							})
-							.option("namespace-id", {
-								type: "string",
-								requiresArg: true,
-								describe: "The id of the namespace to delete from",
-							})
-							.check(demandOneOfOption("binding", "namespace-id"))
-							.option("env", {
-								type: "string",
-								requiresArg: true,
-								describe: "Perform on a specific environment",
-								alias: "e",
-							})
-							.option("preview", {
-								type: "boolean",
-								describe: "Interact with a preview namespace",
-							})
-							.option("force", {
-								type: "boolean",
-								alias: "f",
-								describe: "Do not ask for confirmation before deleting",
-							});
-					},
-					async ({ filename, ...args }) => {
-						await printWranglerBanner();
-						const config = readConfig(args.config as ConfigPath, args);
-						const namespaceId = getKVNamespaceId(args, config);
-
-						if (!args.force) {
-							const result = await confirm(
-								`Are you sure you want to delete all the keys read from "${filename}" from kv-namespace with id "${namespaceId}"?`
-							);
-							if (!result) {
-								logger.log(`Not deleting keys read from "${filename}".`);
-								return;
-							}
-						}
-
-						const content = parseJSON(
-							readFileSync(filename),
-							filename
-						) as string[];
-
-						if (!Array.isArray(content)) {
-							throw new Error(
-								`Unexpected JSON input from "${filename}".\n` +
-									`Expected an array of strings but got:\n${content}`
-							);
-						}
-
-						const errors: string[] = [];
-						for (let i = 0; i < content.length; i++) {
-							const key = content[i];
-							if (typeof key !== "string") {
-								errors.push(
-									`The item at index ${i} is type: "${typeof key}" - ${JSON.stringify(
-										key
-									)}`
-								);
-							}
-						}
-
-						if (errors.length > 0) {
-							throw new Error(
-								`Unexpected JSON input from "${filename}".\n` +
-									`Expected an array of strings.\n` +
-									errors.join("\n")
-							);
-						}
-
-						const accountId = await requireAuth(config);
-
-						await deleteKVBulkKeyValue(accountId, namespaceId, content);
-						await metrics.sendMetricsEvent("delete kv key-values (bulk)", {
-							sendMetrics: config.send_metrics,
-						});
-
-						logger.log("Success!");
-					}
-				);
+		(bulkYargs) => {
+			return kvBulk(bulkYargs.command(subHelp));
 		}
 	);
